pkp/pkp-lib#10571 Move migration into v3.6 folder
