+++ classes/emailTemplate/maps/Schema.php
@@ -84,20 +84,6 @@ class Schema extends \PKP\core\maps\Schema
         $output = [];
 
         $mailableClass = $mailableClass ?? Repo::mailable()->getMailableByEmailTemplate($item);
-        $assignedUserGroupsIds = [];
-        // some mailable are not found during some operations such as performing a search for templates. So ensure mailable exist before using
-        //        if ($mailableClass) {
-        //            $isUserGroupsAssignable = Repo::mailable()->isGroupsAssignableToTemplates($mailableClass);
-        //            $assignedUserGroupsIds = Repo::emailTemplate()->getUserGroupsIdsAssignedToTemplate($item->getData('key'), Application::get()->getRequest()->getContext()->getId());
-        //            if (!$isUserGroupsAssignable) {
-        //                $output['assignedUserGroupIds'] = [];
-        //            } else {
-        //                // Get the current user groups assigned to the template
-        //                $output['assignedUserGroupIds'] = $assignedUserGroupsIds;
-        //            }
-        //        }
-
-
         foreach ($props as $prop) {
             switch ($prop) {
                 case '_href':
